chore: refactor bouncer vault swaps
