--- bouncer/cf-abis
@@ -1,1 +1,0 @@
-../eth-contract-abis/perseverance-rc17
