+++ bouncer/tests/all_concurrent_tests.ts
@@ -16,7 +16,7 @@ import { depositChannelCreation } from './request_swap_deposit_address_with_affi
 import { testDCASwaps } from './DCA_test';
 import { testBrokerLevelScreening } from './broker_level_screening';
 import { testBtcVaultSwap } from './btc_vault_swap';
-import { checkSolEventAccountsClosure } from '../shared/evm_vault_swap';
+import { checkSolEventAccountsClosure } from '../shared/sol_vault_swap';
 
 async function runAllConcurrentTests() {
   // Specify the number of nodes via providing an argument to this script.
